Try to properly handle deallocations
- New version of R-dyntrace and some fixes avoids segfault in GC or
  recursive GC
- Finalized status of functions/vectors should be maintained in one
  place (finalized set in FunctionTable and VectorTable), not two
  (finalized set + finalized flag)
- Keep track of finalized closures, but note that multiple closures may
  be allocated for the same function

There is still some issue where the tracer isn't seeing every
allocation/deallocation, even when the tracer records everything that
happens within the `library` call.

E.g. there are cases where an SEXP address will be reused without
finalizing the old SEXP allocated to that address, or an SEXP will be
finalized but the tracer never saw the allocation.